fix(logger): sort out build
affects: @motech-development/logger
